requirements.txt
Flask
flask-cors
pytesseract
opencv-python
numpy
Pillow
easyocr
paddleocr
werkzeug
